add feature-gate and strict limit for pod template defaults injection (#695)
