+++ pkg/features/kruise_features.go
@@ -69,6 +69,10 @@ const (
 
 	// DaemonWatchingPod enables kruise-daemon to list watch pods that belong to the same node.
 	DaemonWatchingPod featuregate.Feature = "DaemonWatchingPod"
+
+	// PodTemplateNoDefaults to control whether webhook should inject pod default fields into pod template.
+	// If PodTemplateNoDefaults is false, webhook should inject default fields only when the template changed.
+	PodTemplateNoDefaults featuregate.Feature = "PodTemplateNoDefaults"
 )
 
 var defaultFeatureGates = map[featuregate.Feature]featuregate.FeatureSpec{
@@ -83,6 +87,7 @@ var defaultFeatureGates = map[featuregate.Feature]featuregate.FeatureSpec{
 	ResourcesDeletionProtection:      {Default: false, PreRelease: featuregate.Alpha},
 	PodUnavailableBudgetDeleteGate:   {Default: false, PreRelease: featuregate.Alpha},
 	PodUnavailableBudgetUpdateGate:   {Default: false, PreRelease: featuregate.Alpha},
+	PodTemplateNoDefaults:            {Default: false, PreRelease: featuregate.Alpha},
 }
 
 func init() {
