+++ pkg/webhook/imagepulljob/mutating/imagepulljob_create_update_handler.go
@@ -22,6 +22,7 @@ import (
 	"net/http"
 	"reflect"
 
+	"github.com/openkruise/kruise/apis/apps/defaults"
 	appsv1alpha1 "github.com/openkruise/kruise/apis/apps/v1alpha1"
 	"github.com/openkruise/kruise/pkg/util"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -45,7 +46,7 @@ func (h *ImagePullJobCreateUpdateHandler) Handle(ctx context.Context, req admiss
 		return admission.Errored(http.StatusBadRequest, err)
 	}
 	var copy runtime.Object = obj.DeepCopy()
-	appsv1alpha1.SetDefaultsImagePullJob(obj)
+	defaults.SetDefaultsImagePullJob(obj)
 	if reflect.DeepEqual(obj, copy) {
 		return admission.Allowed("")
 	}
